[level 2] Title: 멀리 뛰기, Time: 0.15 ms, Memory: 78.5 MB -BaekjoonHub

+++ 프로그래머스/2/12914. 멀리 뛰기/멀리 뛰기.java
@@ -1,7 +1,8 @@
 class Solution {
     public long solution(int n) {
         if(n == 1) return 1;
-        long[] dp = new long[n+1];
+        
+        int[] dp = new int[n+1];
         dp[1] = 1;
         dp[2] = 2;
